견적현황및 Entity 수정

+++ farmtest/src/main/java/com/kosta/farm/entity/Farmer.java
@@ -27,7 +27,7 @@ import lombok.NoArgsConstructor;
 public class Farmer {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer farmerId;
+	private Long farmerId;
   // FK
   // @Column
   // private Integer userId;
@@ -44,7 +44,7 @@ public class Farmer {
 	@Column
 	private String farmAccountNum; // 팜 계좌번호
 	@Column
-	private String farmInterest1; // 
+	private String farmInterest1; // 관심 농산물
 	@Column
 	private String farmInterest2;
 	@Column
@@ -57,5 +57,5 @@ public class Farmer {
 	@CreationTimestamp
 	private Timestamp createDate;
 	@Column
-	private boolean state;
+	private boolean state; // 
 }
